Backend update
Update the backend for routes. Broke up from a monolithic file into individual pieces.
