--- server/db/index.js
@@ -1,335 +1,0 @@
-// === BACKEND: Secure Express Server ===
-require('dotenv').config();
-const express = require('express');
-const cors = require('cors');
-const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
-const stripe = require('stripe')(process.env.STRIPE_SECRET_KEY);
-const { Pool } = require('pg');
-const { body, validationResult } = require('express-validator');
-
-const app = express();
-
-// Security Middleware
-app.use(helmet());
-app.use(cors({
-  origin: process.env.FRONTEND_URL || 'http://localhost:5173',
-  credentials: true
-}));
-app.use(express.json());
-
-// Rate limiting for auth routes
-const authLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 100 // limit each IP to 100 requests per windowMs
-});
-app.use('/auth/', authLimiter);
-
-// Database setup with connection error handling
-const pool = new Pool({
-  connectionString: process.env.DATABASE_URL,
-  ssl: process.env.NODE_ENV === 'production' ? { rejectUnauthorized: false } : false
-});
-
-pool.on('error', (err) => {
-  console.error('Unexpected error on idle client', err);
-  process.exit(-1);
-});
-
-// JWT Authentication Middleware
-const authenticateToken = (req, res, next) => {
-  const authHeader = req.headers['authorization'];
-  const token = authHeader && authHeader.split(' ')[1];
-  
-  if (!token) return res.sendStatus(401);
-
-  jwt.verify(token, process.env.JWT_SECRET, (err, user) => {
-    if (err) return res.sendStatus(403);
-    req.user = user;
-    next();
-  });
-};
-
-// Admin check middleware
-const isAdmin = (req, res, next) => {
-  if (req.user && req.user.role === 'admin') {
-    next();
-  } else {
-    res.status(403).json({ error: 'Admin access required' });
-  }
-};
-
-// === Routes ===
-
-// --- Authentication Routes ---
-app.post('/auth/register', 
-  [
-    body('email').isEmail().normalizeEmail(),
-    body('password').isLength({ min: 8 }),
-    body('name').notEmpty().trim().escape()
-  ],
-  async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-
-    const { email, password, name, address } = req.body;
-    
-    try {
-      const hashedPassword = await bcrypt.hash(password, 10);
-      const result = await pool.query(
-        'INSERT INTO users (email, password, name, address, role) VALUES ($1, $2, $3, $4, $5) RETURNING id, email, name, role',
-        [email, hashedPassword, name, address, 'user'] // Default role is 'user'
-      );
-      
-      const user = result.rows[0];
-      const token = jwt.sign(
-        { userId: user.id, role: user.role }, 
-        process.env.JWT_SECRET, 
-        { expiresIn: '1h' }
-      );
-      
-      res.status(201).json({ 
-        user: {
-          id: user.id,
-          email: user.email,
-          name: user.name,
-          role: user.role
-        },
-        token
-      });
-    } catch (err) {
-      if (err.code === '23505') {
-        return res.status(409).json({ error: 'Email already exists' });
-      }
-      console.error(err);
-      res.status(500).json({ error: 'Registration failed' });
-    }
-  }
-);
-
-app.post('/auth/login', 
-  [
-    body('email').isEmail().normalizeEmail(),
-    body('password').notEmpty()
-  ],
-  async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-
-    const { email, password } = req.body;
-    
-    try {
-      const result = await pool.query('SELECT * FROM users WHERE email = $1', [email]);
-      if (result.rows.length === 0) {
-        // Generic error message to prevent user enumeration
-        return res.status(401).json({ error: 'Invalid credentials' });
-      }
-
-      const user = result.rows[0];
-      const validPassword = await bcrypt.compare(password, user.password);
-      
-      if (!validPassword) {
-        return res.status(401).json({ error: 'Invalid credentials' });
-      }
-
-      const token = jwt.sign(
-        { userId: user.id, role: user.role }, 
-        process.env.JWT_SECRET, 
-        { expiresIn: '1h' }
-      );
-
-      res.json({ 
-        user: {
-          id: user.id,
-          email: user.email,
-          name: user.name,
-          role: user.role
-        },
-        token 
-      });
-    } catch (err) {
-      console.error(err);
-      res.status(500).json({ error: 'Login failed' });
-    }
-  }
-);
-
-app.get('/auth/me', authenticateToken, async (req, res) => {
-  try {
-    const result = await pool.query(
-      'SELECT id, email, name, address, role FROM users WHERE id = $1',
-      [req.user.userId]
-    );
-    if (result.rows.length === 0) {
-      return res.sendStatus(404);
-    }
-    res.json(result.rows[0]);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: 'Failed to fetch user data' });
-  }
-});
-
-// --- Product Routes ---
-app.get('/products', async (req, res) => {
-  try {
-    const result = await pool.query('SELECT * FROM products WHERE stock > 0');
-    res.json(result.rows);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: 'Failed to fetch products' });
-  }
-});
-
-app.get('/products/:id', async (req, res) => {
-  try {
-    const result = await pool.query('SELECT * FROM products WHERE id = $1', [req.params.id]);
-    if (result.rows.length === 0) {
-      return res.status(404).json({ error: 'Product not found' });
-    }
-    res.json(result.rows[0]);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: 'Failed to fetch product' });
-  }
-});
-
-// Protected product management routes
-app.post('/products', authenticateToken, isAdmin, 
-  [
-    body('artist').notEmpty().trim().escape(),
-    body('description').notEmpty().trim().escape(),
-    body('price').isFloat({ min: 0.01 }),
-    body('stock').isInt({ min: 0 })
-  ],
-  async (req, res) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({ errors: errors.array() });
-    }
-
-    const { artist, description, price, image, genre, stock } = req.body;
-    
-    try {
-      const result = await pool.query(
-        'INSERT INTO products (artist, description, price, image, genre, stock) VALUES ($1, $2, $3, $4, $5, $6) RETURNING *',
-        [artist, description, price, image, genre, stock]
-      );
-      res.status(201).json(result.rows[0]);
-    } catch (err) {
-      console.error(err);
-      res.status(500).json({ error: 'Failed to create product' });
-    }
-  }
-);
-
-// Similar improvements for PUT and DELETE product routes...
-
-// --- Cart Routes ---
-app.get('/cart', authenticateToken, async (req, res) => {
-  try {
-    const result = await pool.query(
-      `SELECT c.id, p.id as product_id, p.artist, p.price, p.image, c.quantity 
-       FROM cart_items c
-       JOIN products p ON c.product_id = p.id
-       WHERE c.user_id = $1`,
-      [req.user.userId]
-    );
-    res.json(result.rows);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ error: 'Failed to fetch cart' });
-  }
-});
-
-// Similar improvements for other cart routes...
-
-// --- Order Routes ---
-app.post('/orders', authenticateToken, async (req, res) => {
-  const client = await pool.connect();
-  
-  try {
-    await client.query('BEGIN');
-    
-    // 1. Get user's cart items
-    const cartItems = await client.query(
-      `SELECT c.product_id, c.quantity, p.price, p.stock 
-       FROM cart_items c
-       JOIN products p ON c.product_id = p.id
-       WHERE c.user_id = $1`,
-      [req.user.userId]
-    );
-    
-    if (cartItems.rows.length === 0) {
-      return res.status(400).json({ error: 'Cart is empty' });
-    }
-    
-    // 2. Check stock and calculate total
-    let total = 0;
-    for (const item of cartItems.rows) {
-      if (item.quantity > item.stock) {
-        throw new Error(`Insufficient stock for product ${item.product_id}`);
-      }
-      total += item.quantity * item.price;
-    }
-    
-    // 3. Create order
-    const orderResult = await client.query(
-      'INSERT INTO orders (user_id, total, status) VALUES ($1, $2, $3) RETURNING id',
-      [req.user.userId, total, 'pending']
-    );
-    const orderId = orderResult.rows[0].id;
-    
-    // 4. Create order items and update product stock
-    for (const item of cartItems.rows) {
-      await client.query(
-        'INSERT INTO order_items (order_id, product_id, quantity, price) VALUES ($1, $2, $3, $4)',
-        [orderId, item.product_id, item.quantity, item.price]
-      );
-      
-      await client.query(
-        'UPDATE products SET stock = stock - $1 WHERE id = $2',
-        [item.quantity, item.product_id]
-      );
-    }
-    
-    // 5. Clear cart
-    await client.query(
-      'DELETE FROM cart_items WHERE user_id = $1',
-      [req.user.userId]
-    );
-    
-    await client.query('COMMIT');
-    
-    res.status(201).json({ 
-      message: 'Order created successfully',
-      orderId
-    });
-  } catch (err) {
-    await client.query('ROLLBACK');
-    console.error(err);
-    res.status(400).json({ error: err.message });
-  } finally {
-    client.release();
-  }
-});
-
-// Server startup with env validation
-const requiredEnvVars = ['DATABASE_URL', 'JWT_SECRET', 'STRIPE_SECRET_KEY'];
-const missingVars = requiredEnvVars.filter(varName => !process.env[varName]);
-
-if (missingVars.length > 0) {
-  console.error(`Missing required environment variables: ${missingVars.join(', ')}`);
-  process.exit(1);
-}
-
-const PORT = process.env.PORT || 4242;
-app.listen(PORT, () => {
-  console.log(`Server running on port ${PORT}`);
-});
