+++ server/middleware/isAdminMiddleware.js
@@ -1,0 +1,9 @@
+const isAdmin = (req, res, next) => {
+  if (req.user && req.user.role === 'admin') {
+    next();
+  } else {
+    res.status(403).json({ error: 'Admin access required' });
+  }
+};
+
+module.exports = isAdmin;
